+++ src/patkit/__init__.py
@@ -56,7 +56,7 @@ from .cli import run_cli
 from .initialise import add_derived_data, initialise_patkit
 from .qt_annotator import run_annotator
 
-# __all__ = ['add_derived_data', 'initialise_patkit']
+__all__ = ['add_derived_data', 'initialise_patkit', 'run_cli', 'run_annotator']
 
 # Load logging config from json file.
 LOG_CONFIG = "patkit_logging_configuration.json"