+++ docs/Changelog.markdown
@@ -39,11 +39,26 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
   - The alpha and beta versions will be mainly quality assurance and code clean up.
 
 
-<!-- 
+## [0.22.1] - 2026-05-23
+
+### Highlights
+
 - Manifest files now work.
 - Some corrections to documentation.
   - Links on PyPi will now hopefully work. 
--->
+
+### Added
+
+- First automated unit tests and integration tests.
+
+### Bugs
+
+- [Old Bugs](Bugs.markdown).
+
+### Fixed
+
+- Manifest overwrite bug.
+- Links on PyPi
 
 
 ## [0.22.0] - 2026-04-14